requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==0.19.0
requests==2.26.0
together==1.5.5
gunicorn==23.0.0
